Create intro
The Community "Building Immune Digital Twins" aims to make immune digital twin technology a reality. Our main purpose is to bring together researchers across disciplines for activities ranging from extended active teamwork on specific immune digital twin projects, webinars, new collaborative projects, and applications.   

More specifically: 
We aim to create an open-access repository of existing models and develop technology that enables models to be integrated, re-used, adapted, or expanded to serve the purpose of building Immune Digital Twins (IDTs) across scales. 

We aim to provide a guide of best practices for the development of a FAIR environment for the building, hosting, and deployment of IDTs.  Develop standards for Digital Twin technology to facilitate the reusability and integration of computational models across scales, but also to facilitate the technology development that would need to implement these standards in order to maximize interoperability. 

We aim to foster a network of collaborators and experts in all relevant areas of research. The ultimate goal is to help create a long-term interdisciplinary immune digital twin community, willing to take up the challenges of this exciting new field.